+++ CHANGELOG.md
@@ -67,6 +67,8 @@ Modified clusters:
 - [matter.js]: Update to 0.13.0-alpha.0-20250415-475996bb5.
 - [matter.js]: Update to 0.13.0-alpha.0-20250418-8cfc0b832.
 - [matter.js]: Update to 0.13.0-alpha.0-20250420-9f45e4f77.
+- [matter.js]: Update to 0.13.0-alpha.0-20250420-9f45e4f77.
+- [matter.js]: Update to 0.13.0-alpha.0-20250423-8917d1d1d.
 - [help]: Updated cli help screen.
 - [logger]: Improved frontend logger cleaning.